+++ src/renderer/src/features/action/run/index.ts
@@ -1,0 +1,3 @@
+export { runAction } from './model/runAction';
+export { useActionRuns } from './model/useActionRuns';
+export type { ActionRun } from './model/types';